Replaced old sync_dependencies.py with sync_dependencies_new.py, and disable CI
